#196, io writers should check destination more thoroughly before deleting

+++ src/cogent3/app/data_store.py
@@ -566,11 +566,24 @@ class WritableDirectoryDataStore(ReadOnlyDirectoryDataStore, WritableDataStoreBa
         self._persistent = {k: v for k, v in d.items() if k != "self"}
         self.mode = mode
 
+    def _has_other_suffixes(self, path, suffix):
+        p = Path(path)
+        for f in p.iterdir():
+            if get_format_suffixes(str(f))[0] != suffix:
+                return True
+        return False
+
     def _source_create_delete(self, if_exists, create):
         exists = os.path.exists(self.source)
         if exists and if_exists == RAISE:
             raise RuntimeError(f"'{self.source}' exists")
         elif exists and if_exists == OVERWRITE:
+            if self._has_other_suffixes(self.source, self.suffix):
+                raise RuntimeError(
+                    f"Unsafe to delete {self.source} as it contains ",
+                    f"files other than {self.suffix}."
+                    " You will need to remove this directly yourself.",
+                )
             try:
                 shutil.rmtree(self.source)
             except NotADirectoryError:
@@ -635,12 +648,24 @@ class WritableZippedDataStore(ReadOnlyZippedDataStore, WritableDataStoreBase):
         self._persistent = {k: v for k, v in d.items() if k != "self"}
         self.mode = "a" or mode
 
+    def _has_other_suffixes(self, path, suffix):
+        for f in zipfile.ZipFile(path).namelist():
+            if get_format_suffixes(f)[0] != suffix:
+                return True
+        return False
+
     def _source_create_delete(self, if_exists, create):
         exists = os.path.exists(self.source)
         dirname = os.path.dirname(self.source)
         if exists and if_exists == RAISE:
             raise RuntimeError(f"'{self.source}' exists")
         elif exists and if_exists == OVERWRITE:
+            if self._has_other_suffixes(self.source, self.suffix):
+                raise RuntimeError(
+                    f"Unsafe to delete {self.source} as it contains ",
+                    f"files other than {self.suffix}."
+                    " You will need to remove this directly yourself.",
+                )
             os.remove(self.source)
         elif dirname and not os.path.exists(dirname) and not create:
             raise RuntimeError(f"'{dirname}' does not exist")
